fix
duplications sonar qube
